refactor: feignclient 패키지 구조 수정 및 네이버 api error 추가
